Refactored more extractors
Refactored the SourceMonitor, the CK, the MOOD and the Halstead.
A few methods are still missing, generated from the methods_per_file.
From this only need to implement the methods_per_file and Designite.
In Designite, need to refactor and implement the organic processor.
In the end, need to create cache for methods_per_file.
Moreover, need to include the serializer for the data and the
accessors for data queries.

Signed-off-by: Bruno Sotto-Mayor